requirements.txt
python>=3.9
matplotlib
numpy
optuna=3.2.0
pandas
pot=0.9.0
pymysql
scikit-learn
scipy
seaborn
sqlalchemy
sqlalchemy-utils
pytorch=2.0.1
tqdm
umap-learn